requirements.txt
pydantic~=2.10
httpx~=0.24.1
requests~=2.32.3
sqlalchemy~=2.0.22
psycopg2-binary~=2.9.6
apscheduler~=3.11.0
schedule~=1.2.2
python-dotenv~=1.0.0
fastapi~=0.110.0
uvicorn~=0.29.0
playwright~=1.41.1
beautifulsoup4~=4.12.2
